requirements.txt
Flask
tensorflow
keras
keras-tcn
pymysql
numpy
